* src/GlueXUserEventInformation.cc, src/GlueXUserEventInformation.hh [rtj]    - remove constructors with explicit primary particle specs in the      arguments because multiple particles are generally specified to      launch an event.    - introduce new methods AddBeamParticle, AddTargetParticle, and      AddPrimaryVertex to add beam, target, and final-state particles      to the MC record for the event.    - add support for the new polarization tag
* src/GlueXPrimaryGeneratorAction.cc, src/GlueXPrimaryGeneratorAction.hh [rtj]
   - changes to support the modified interface to GlueXUserEventInformation
     class objects described above
   - fix a bug in the ImportanceSampling class, where int class
     variables Npassed, Nfailed were doomed to overflow into negative values,
     changed to long int type.
   - rearranged the formulas for Scoherent, Sincoherent to prevent
     errors from overflow under multiplication of two large ints.

+++ src/GlueXPrimaryGeneratorAction.hh
@@ -167,8 +167,8 @@ class GlueXPrimaryGeneratorAction : public G4VUserPrimaryGeneratorAction
       double Psum;
       double Pcut;
       double Pmax;
-      int Nfailed;
-      int Npassed;
+      long int Nfailed;
+      long int Npassed;
 
       ImportanceSampler()
        : Psum(0), Pcut(1), Pmax(0), Nfailed(0), Npassed(0) {}
